Add typing to the method return and remove unncessary data parameter conditional
