+++ server/api/discord.ts
@@ -1,4 +1,4 @@
-import axios from 'axios';
+import axios, { AxiosPromise, AxiosResponse } from 'axios';
 
 import {
   clientId,
@@ -7,16 +7,21 @@ import {
   BASE_DISCORD_API_URL
 } from '../src/config';
 
+interface DiscordOAuthTokenResponseData {
+  token_type: string
+  access_token: string
+}
 
+interface DiscordUserData {
+  username: string
+  discriminator: string
+}
 
 export class Discord {
 
   /** Axios request builder */
-  static async request(endpoint: string, headers: any, body = {}, method = "GET") {
+  static async request(endpoint: string, headers: any, data: URLSearchParams = undefined, method = "GET") {
     const url = `${BASE_DISCORD_API_URL}/${endpoint}`;
-    const data = (method === "GET")
-      ? undefined
-      : body;
 
     const response = (await axios({ method, url, data, headers })).data;
     return response;
@@ -25,7 +30,7 @@ export class Discord {
   // Individual Discord API routes
 
   /** Get an access token from Discord OAuth API */
-  static async getDiscordToken(authCode: string) {
+  static async getDiscordToken(authCode: string): Promise<DiscordOAuthTokenResponseData> {
     const params = new URLSearchParams({
       client_id: clientId,
       client_secret: clientSecret,
@@ -42,7 +47,7 @@ export class Discord {
   }
 
   /** Get a user object from Discord users API  */
-  static async getDiscordUser(tokenType: string, accessToken: string) {
+  static async getDiscordUser(tokenType: string, accessToken: string): Promise<DiscordUserData> {
     const headers = { authorization: `${tokenType} ${accessToken}` };
 
     const res = await this.request('users/@me', headers);
